+++ apps/menu-bar/electron/src/TrayGenerator.ts
@@ -7,6 +7,8 @@ import {
   ipcMain,
   app,
   nativeTheme,
+  Rectangle,
+  Display,
 } from 'electron';
 import path from 'path';
 
@@ -20,25 +22,53 @@ export default class TrayGenerator {
     this.tray = null;
     this.mainWindow = mainWindow;
   }
-  getWindowPosition = () => {
+  calculateWindowPosition = (display: Display) => {
     if (this.tray === null) {
       return;
     }
     const windowBounds = this.mainWindow.getBounds();
     const trayBounds = this.tray.getBounds();
+    const { workArea } = display;
+    const PADDING = 12;
 
-    const primaryDisplay = screen.getPrimaryDisplay();
-    const { width } = primaryDisplay.workAreaSize;
+    // Determine closest horizontal edge
+    const trayCenterX = trayBounds.x + trayBounds.width / 2;
+    const distanceToLeft = trayCenterX - workArea.x;
+    const distanceToRight = workArea.x + workArea.width - trayCenterX;
+    let isLeft = distanceToLeft < distanceToRight;
 
-    const x = Math.round(width - windowBounds.width - 12);
-    const y = Math.round(trayBounds.y - windowBounds.height - 12);
+    // Determine closest vertical edge
+    const trayCenterY = trayBounds.y + trayBounds.height / 2;
+    const distanceToTop = trayCenterY - workArea.y;
+    const distanceToBottom = workArea.y + workArea.height - trayCenterY;
+    let isTop = distanceToTop < distanceToBottom;
+
+    const zeroTray =
+      trayBounds.x === 0 && trayBounds.y === 0 && trayBounds.width === 0 && trayBounds.height === 0;
+    if (process.platform === 'linux' && zeroTray) {
+      // On Linux trayBounds may return an empty Rectagle, in these cases force to top-right corner
+      isTop = true;
+      isLeft = false;
+    }
+
+    const x = isLeft
+      ? workArea.x + PADDING
+      : workArea.x + workArea.width - windowBounds.width - PADDING;
+
+    const y = isTop
+      ? workArea.y + PADDING
+      : workArea.y + workArea.height - windowBounds.height - PADDING;
+
     return { x, y };
   };
-  showWindow = () => {
-    const primaryDisplay = screen.getPrimaryDisplay();
-    const { height, width } = primaryDisplay.size;
+  showWindow = (bounds?: Rectangle) => {
+    const currentDisplay = bounds
+      ? screen.getDisplayMatching(bounds)
+      : screen.getDisplayNearestPoint(screen.getCursorScreenPoint());
+
+    const { height, width } = currentDisplay.size;
     this.mainWindow.webContents.send('popoverFocused', { screenSize: { height, width } });
-    const position = this.getWindowPosition();
+    const position = this.calculateWindowPosition(currentDisplay);
 
     if (!position) {
       return;
@@ -49,11 +79,11 @@ export default class TrayGenerator {
   hideWindow = () => {
     this.mainWindow.hide();
   };
-  toggleWindow = () => {
+  toggleWindow = (_: KeyboardEvent, bounds: Rectangle) => {
     if (this.mainWindow.isVisible()) {
       this.hideWindow();
     } else {
-      this.showWindow();
+      this.showWindow(bounds);
     }
   };
   rightClickMenu = () => {
@@ -89,11 +119,11 @@ export default class TrayGenerator {
     this.tray.on('click', this.toggleWindow);
     this.tray.on('right-click', this.rightClickMenu);
 
-    ipcMain.handle('open-popover', this.showWindow);
+    ipcMain.handle('open-popover', () => this.showWindow());
     ipcMain.handle('close-popover', this.hideWindow);
 
-    app.on('open-url', this.showWindow);
-    app.on('second-instance', this.showWindow);
+    app.on('open-url', () => this.showWindow());
+    app.on('second-instance', () => this.showWindow());
 
     this.mainWindow.on('blur', () => {
       if (!this.tray) {