+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### 🐛 Bug fixes
 
 - Fix cache check to return app path instead of directory. ([#266](https://github.com/expo/orbit/pull/266) by [@altaywtf](https://github.com/altaywtf))
+- [electron] Fix multi display popover position logic. ([#265](https://github.com/expo/orbit/pull/265) by [@gabrieldonadel](https://github.com/gabrieldonadel))
 
 ### 💡 Others